Add support for deleting tables, and creating, deleting, querying and testing the existence of datasets. In addition, this makes the table name parser a static public method and makes it more general, and takes advantage of that in several table-related APIs. Finally, this adds a sample notebook making use of some of these APIs and illustrating populating a BQ table from a Pandas dataframe.
Some rearrangement of the APIs. exists() and delete() now live on the objects to which they apply
while object creation is handled one level up (e.g Table inserts by TableList.insert, dataset inserts
by bq.tables with an additional flag).

Also, the table schema is now treated more consistenly as a TableSchema object (although internally
we save the Python list representation so we can use it for table creation without having to regenerate
it).

Don't check if file exists before deletion, just handle the exception.

Revert an unintentional change.

Move create() back to be instance methods of Table and DataSet.

Fix the representation of a schema in the stringized notebook.

Fix two issues that were breaking the existing notebooks.

- don't enforce the project ID having to be the default
- make sure query timeouts are integer values

Change the __repr__ for Table to be the empty string.

Add section headers and a better intro to the Pandas import notebook.

Various modifications to sample notebook

Add __repr__ for DataSet