Implementation of Views.
Unlike the previous version, in this version a View wraps a Table (actually
BaseTable) and a Query (for metrializing the view). That allows it to support
a mix of Table and Query functionality through thin wrappers.

As part of this change the subset of Table functionality that is relevant to
View has been factored out into a BaseTable class.

Sample useage:

  v = bq.view(<name>)
  v.create(<query>)
  v.results().to_dataframe()

  v.update(query=<query>)
  v.sample(count=10)
  v.delete()