Handle parsing data for tables with nested RECORD fields.
For now if the nested field is REPEATING we only handle the first
value in the underlying array of subfields. We could handle more and
add .0, .1, etc suffixes on the names, but this may have its own issues.

Note too that we only handle currently REPEATING in the case of RECORD
fields, not scalar fields.

In order to support this, we need to use the underlying _bq_schema in
Schema. This is happening in more and more of the code, leading me to
question how much value the flattened Schema object is providing, but
that is not something I want to address yet.