Modify the server/UX build to be per-file. Adjust the references accordingly so the build still works. Use symlinks for the common file directory.
The purpose of this is to allow the use of Eclipse with the Angular and TypeScript
plugins. Once this is done it is easy to make a DataLabNode project and a DataLabUI
project in Eclipse for these two projects, and we get autocompletion and other IDE
goodness. We can build the DataLabUI project from within Eclipse too; there are still a
couple of issues with building the DataLabNode project; see:

    https://github.com/axmor/typecs/issues/45

and the much less blocking issue:

    https://github.com/axmor/typecs/issues/44

I'm hoping that we would get in-IDE debugging too although I haven't tested that yet as
the server build is not 100% complete.

This change does not include the Eclipse project files themselves; we may not want to
include those in the repo (we can and should still ne an;e to build with build.sh).
It does not update .gitignore yet to exclude these either; will do that separately
if/when this is merged.

build.sh still builds files together; I added a new script to validate building
them separately.