Improved handling of nested tables.
This changes the parser for table results so that it can handle nested table
results, which it turns into hierarchical Python structures (that could be
usefully JSON-encoded, for example). It also changes the chart/table
magic code so that we only return the top-level non-RECORD fields so that we
can at least partly view nested tables.

It would be good to improve this to flatten the tables if there are RECORDs
that contain non-repeating fields. We could also flatten the first level of
repeating fields in tables by turning them into HTML lists. However, these
can be done later/separately.