Add a script to create symlinks and tsconfig.json files to work with Typescript IDEs like VS Code and Atom.
I tried to make the tsconfig files work with the shared directory to create a shared library but
couldn't get this to work. Creating the symlinks still seems the best approach for now until the
tools are more flexible. The script will create the tsconfig files too which are treated as project
files by some IDEs for TypeScript (e.g. Atom with the TypeStrong plugin works very well). You can also
compile the code now by just running tsc with no arguments in the directory that has the tsconfig file
(or in one of its subdirectories).

I'm not changing build.sh at all because the tsconfig approach still relies on the symlinks so it
isn't portable.