Build DataLab on Jupyter
- Install and launch Jupyter with Jupyter-specific config
- Updated proxy server to use custom HTML content for UX
- Updated custom script and stylesheet for Jupyter
- New WebSocket shimming

- Temporarily work against sample notebooks exclusively
This is until we work out git integration/workspace. For both cloud and
local runs, the directory listing will go against just the sample
notebooks, either those within the container (in case of cloud) or those
mounted into the container (in case of local).

- Force-trust all notebooks
This is done on the client by marking content loaded from the server as
trusted.

- Add support for downloading and converting notebooks to HTML/python.
- Print stylesheet + getting DataLab to work in Safari

- Analytics logging w. user hash and project number
- Updated notebooks to v4 format