Add validate/execute/deploy actions to bigquery pipeline and change variable specs to be JSON/YAML.
Change config blocks for cell magics to be JSON or YAML rather than Python code.

For now I am not making variables for in an 'args' subsection of the config block.
It may be a good idea to do this to ensure they are namespaced separately from other
config values, but right now there are no other config values, and arguably it may be
cleaner to put other config values in a subsection rather than the other way around.