Low-fi version check.
This adds a version check that can show a custom message in the info modal
dialog. The message can be for either a recommended or required update.
The versions and messages are controlled from a file in GCS; currently
gs://cloud-datalab/version.json, which needs to be publicly accessible via
HTTPS. The fill should contain something like:

jsonp({
  "required": {
    "version": "1.1",
    "message": "Upgrade required!"
  },
  "recommended": {
    "version": "1.2",
    "message": "Version 1.2 is now available. An upgrade is recommended."
  }
})

The appropriate message is added to the info dialog HTML under the preformatted
version section.

No attempt to block functionality is being done, and this is only surfaced in the
info dialog at present.

I'm not sure if we need something more sophisticated than this; just opening it
up for discussion.