Fix the storage command sample notebook and fix a bug in list expansion in storage magics.
This notebook was not in the automated notebook tests as originally it was so
far out of date the automated test failed on (almost?) every cell as all the
new outputs were different. Unfortunately I didn't remember that it had been
skipped so when testing the notebook updates recently it was overlooked.