Add (back) MathJax support.
We have had two customers request this so far via Google-Cloud-Datalab-Feedback@.

I looked at the size of the scripts that get downloaded due to this. While the
MathJax source is about 17MB, the minified scripts are not bad at all, coming in
at about 400kB.

I went for the CDN-hosted scripts as these may already be cached and may be faster
to retrieve than through hosting in the container. That said, I went for the the
2.5 release; we could instead just use 'latest' in the URL to get the latest
release which could improve likelihood of cached versions.

As far as the impact supporting MathJax could have on a new UX is concerned, it
can be estimated by looking at the impact on Jupyter, which is mostly in:

  https://github.com/jupyter/notebook/blob/master/notebook/static/notebook/js/mathjaxutils.js

textcell.js and outputarea.js each have a few lines referencing this; most other
MathJax-related code in Jupyter (including some of mathjaxutils.js) is related just
to MathJax-related config (enabled or not, URL value, and handling load failure).

Issue #681