A simple alternative to widgets.
One of the complaints we have had from users is the lack of support for Jupyter widgets.
This change addresses some of that limitation by allowing the variables in %%sql modules
that are displayed with %%chart to include simple dynamic controls which will re-execute
the query with the new values when changed (with client-side caching of past results).

This is controlled via the YAML chart options, in a 'variables' section. There are five
types of controls supported, 'text' (textbox), 'set' (checkbox group), 'picker' (dropdown
allowing a single value), 'slider' (combined slider/textbox for numeric value with
min/max/step parameters), and 'boolean' (a checkbox for a Boolean value).

 Each can have an optional 'label' property that defaults to the variable name, and a
'value' property to set the initial value, which, if not present, defaults to the default
value in the %%sql cell. The control type is set by a 'type' property which will default to
'text' for strings and 'set' for lists but must otherwise be specified.

For sliders, if s 'step' property is not specified and 'min' and 'max' are ints a 'step' of
1 will be used; else a step of (max - min) / 10.0 will be used.

For pickers, if the variable is not a list, a list of values can be provided with the
'choices' property.

Some examples:

    %%sql -m mother_ages

    state=['WA', 'TX', 'AZ']

    SELECT mother_age, COUNT(*) AS count FROM [publicdata:samples.natality]
    WHERE state IN $state
    GROUP BY mother_age
    ORDER BY mother_age

    %chart paged_table -d mother_ages
    variables:
      state:
        label: 'State'

The above will create a set of three checkboxes which can be checked/unchecked in different
combinations to affect the resulting table via altering the IN clause of the query.

    %%sql -m total_for_age_by_state

    age=20

    SELECT state, COUNT(*) AS count FROM [publicdata:samples.natality]
    WHERE mother_age==$age
    GROUP BY state
    ORDER BY state

    %chart table -d total_for_age_by_state
    variables:
      age:
        type: slider
        min: 10
        max: 60

The above will create a slider to allow an age in the range 10 through 60 to be selected.

    %%sql -m total_by_state
    state = 'WA'
    married = True

    SELECT mother_age, COUNT(*) AS count FROM [publicdata:samples.natality]
    WHERE state==$state AND mother_married==$married
    GROUP BY mother_age
    ORDER BY mother_age

    %chart columns -d natality
    title: Births by Age and Marital Status of Mother
    hAxis:
      title: Age
    variables:
      state:
        type: picker
        choices: [AZ, TX, WA]
      married:
        type: boolean

The above will create both a picker for the three choices of state and a checkbox to toggle
the value of martial status.