Monitoring API
This initial version of the monitoring client for Datalab allows users
to query timeseries data. It introduces a new module under gcp that
users can import using:

    from gcp.stackdriver import monitoring

Querying timeseries data:

There is a `Query` class that allows users to query timeseries data
for their monitored resources. They can initialize the query by specifying a
metric type and time interval, and refine it by adding filters to it.

The timeseries data is returned as a pandas DataFrame that allows users
to further manipulate the data and visualize it within Datalab.

IPython magics:

There are a couple of IPython magic commands to allow users to list
details of the available metrics and resource types. E.g.:

    %%monitoring list metrics

The base library:

The _impl directory is a snapshot of an earlier version of a
monitoring client library that we have submitted to gcloud-python:

https://github.com/GoogleCloudPlatform/gcloud-python/pull/1691

It works as the base library on top of which the Datalab library adds
interactive features.

The authors of this code are @rimey and myself, both at Google.