Fix two issues that prevented logging in from a local instance.
The first issue was around the creation of the .config directory.
We had code in multiple places that tried to create it, and in one
of those cases it tried to create that directory repeatedly, resulting
in failure messages in the logs after the first attempt.

This change fixes that by making sure that only the EULA accepted
handler creates that directory, and only in the case that it does
not already exist.

The other issue was the OAuth callback handler, which did not make
sure the OAuth client instance existed before attempting to use it.

This change fixes that by adding that check.

This also adds (minimum) support for running on a port other than 8081.