Add the sign-in/sign-out buttons to the notebooks UI.
There are some sample notebooks which require credentials to run
(because they call into external APIs). If you try to run those
notebooks while signed out, then they will produce an error
message helpfully suggesting that you sign.

However, the sign-in button was only being shown on the tree-view
page and not on the notebook-view page.

This meant that you had to navigate back to the tree-view page,
sign in, and then go back to the notebook to retry running the
code.

This change fixes that by adding the sign-in/sign-out buttons
to the notebook-view page in the exact same place they appear
for the tree-view page.