Add a new template for the file editing page.
We previously were serving the file editing page using the
built-in Jupyter UI. That meant that it looked wildly different
from the rest of Datalab, and was missing the Datalab-specific
UI elements like the sign-in and feedback buttons.

This version of the editor page is functional, but not yet
feature complete. The remaining missing features will be added
in later commits.