Make the file editor download link support query parameters.
This change makes it so that the file download link in the
editor UI will still work if the UI has been visited with a
URL containing query parameters (e.g. `?foo=bar`).

However, note that any such parameters are not preserved. The
download link does not support any query parameters other than
`?download`, so any other parameters are discarded.