Add support for connecting to a kernel gateway.
This change allows Datalab to use a
[kernel gateway](https://github.com/jupyter/kernel_gateway) for
running the notebook kernels.

The setting for whether or not to use a gateway is based on the
`KG_URL` environment variable, which defines the URL of the
gateway (if any) to use.

If that environment variable is set, then *all* kernels are run
using that gateway. Otherwise, all kernels are run locally within
the Datalab container.

For example:

    EXPERIMENTAL_KERNEL_GATEWAY_URL=http://localhost:8080 ./containers/datalab/run.sh

will run Datalab with all kernels being run by a kernel gateway
running on localhost:8080.

... whereas:

    ./containers/datalab/run.sh

will run Datalab will all kernels running inside of the Datalab
container.

Note: This uses the
[nb2kg server extension](https://github.com/jupyter/kernel_gateway_demos/tree/master/nb2kg),
which is demo code. As such, we require that the user explicitly opt-in
to using it.