Make the Datalab server kill abandoned Jupyter servers.
This change makes the Datalab NodeJS server, at startup, list
all processes for the 'jupyter-notebook' binary, and kill them.

This fixes #892, where those processes could become abandoned
if the NodeJS server gets restarted.

This should be a safe thing to do at startup, as the NodeJS
server only runs inside of a Docker container, and so it has
an isolated process space.

I.E. the only processes that should be killed are those that
were started by an earlier incarnation of the NodeJS server.