Split the `sessions` UI out into its own page.
Previously, the `tree` UI included two unrelated views; a directory
view of the file system, and a list view of running sessions.

This was confusing as it was not intuitive that you should open
the file listing in order to be able to switch to the session
list, nor was it obvious that the UI controls on that page
toggled the type of view that was presented.

This change tries to clean all of that up by splitting the
session listing into its own page, and putting a link to that
page in the button toolbar that is in the top, right-hand
corner of the UI.