Remove potentially sensitive information from the output of '/_info'
This change scrubs the request headers that are included in the
output of the '/_info' handler so that any fields that might
include auth credentials are deleted before being returned to
the user.

That, in turn, makes it less likely for users to accidentially
leak their personal credentials if they try to include the output
of the '/_info' handler when submitting a bug report.