Initial image for the kernel-gateway
This change takes the existing, main Dockerfile for Datalab, and
splits it into three separate images:

1. A base image used by the other images. This includes the
   pydatalab library and it's assorted run-time dependencies.
2. The main image for running a stand-alone Datalab instance.
3. A new image for running a kernel gateway for Datalab.

Each of the three images also was given a corresponding `README.md`
file describing them.

Although the main image remains usable by itself, with this change
it will be possible to run the Datalab UI on a separate machine
from the one used for running the kernels.

Additional instructions for running Datalab in such a split
configuration will be coming in a later change.