Add a new Docker image for running with a backend on GCP.
This encapsulates all of the logic for spinning up a GCE
VM running a Datalab kernel gateway, setting up an SSH
tunnel connected to that VM, and then running a local
Datalab instance pointed at that SSH tunnel.

This should make it much simpler to run Datalab with kernels in the cloud.