Preliminary usage reporting.
This change adds anonymous reporting of usage statistics that
can be used to inform future development work on Datalab.

The actual reporting code is relatively simple, but this change
is large because it also introduces two ways to disable
that reporting:

1. A new "ENABLE_USAGE_REPORTING" environment variable has been
   added to the Docker image. This defaults to 'true', but can
   be overridden at the command line by adding an
   '-e ENABLE_USAGE_REPORTING=false' argument to the Docker
   command used to run Datalab. Setting this to false disables
   all usage reporting.
2. A new concept of user settings is added to the product.
   These are stored in the file 'datalab/.config/settings.json'
   and can be updated by either manually updating that file,
   or by calling into a new '/setting?key=<key>&value=<value>'
   handler that this change introduces.

This change also introduces a new static page at
'/static/reporting.html', that both describes why the tool is
reporting usage statistics, and provides a form to enable/disable
that reporting.

Finally, this change also incorporates updates to the contents
and serving of the EULA page to provide notice of the statistics
reporting:

1. Change the EULA text to describe the collection being performed
   and how to disable it.
2. Allow the EULA page to reference static resources, so that
   we can display the icon that the user needs to press to turn
   off the collection.
3. Change the directory that we look for to see if the EULA has
   been agreed to from the generic '.config' directory (which
   gcloud also writes) to the eula-specific '.config/eula'
   directory).

NOTE: If *either* the environment variable or the user setting
is used to disable reporting of usage statistics, then no reporting
will happen. Reports are only sent when *both* the environment
variable allows it, *and* the user settings allow it.