Remove the need to login from the command line
This change allows the user to perform the OAuth flow to log in
to Datalab at startup via the web browser.

That, in turn, streamlines the user experience for running Datalab
with a backend in GCP. This is because the user may not be logged
in initially, but we need them to be logged in before we can create
an SSH tunnel connected to their VM.

Previously, we satisfied that requirement by using the 'gcloud auth
login' flow from the command line, but that is cumbersome. With this
change that flow is now performed using the browser, which makes
it less cumbersome.

This change also makes the `setGcloudAccount` method synchronous.

Although NodeJS code is normally asynchronous, the logic for
persisting user credentials (the `persistCredentials` method) needs to
successfully complete before subsequent code in the same request handler
runs, in case that code needs to use the persisted credentials.

Accordingly, the rest of the code for handling that part of the login
process was synchronous, but the call to `gcloud config set account`
was not.

This change fixes that so that the entire `persistCredentials`
method is synchronous; once it completes the caller can rely
on the credentials having been successfully persisted.