Initial support for custom themes
This adds a new user setting called 'theme' that can be set to
one of 'light', 'dark', or 'custom'.

Both 'light' and 'dark' are canned themes with 'light' matching the
existing Datalab style, and 'dark' modifying that theme to have
dark backgrounds rather than light ones.

The 'custom' setting is for using a user-specified CSS file stored
in the 'datalab/.config' directory named 'custom.css'.

This change does not incorporate a UI for modifying the theme
setting, but the user can change it by directly modifying the
'datalab/.config/settings.json' file.

The canned dark theme still needs some work.

This also adds cache-invalidation for the static files.

This commit adds a distinction between static files that are
really static (i.e. part of the Docker container) and those that
are pseudo-static (part of the user's mounted volume and potentially
updated).

It then adds cache-invalidation for those potentially updated
files based on watching the file for changes and clearing the
cache when one is detected.

For now, the only such file is the user's custom theme.