Create the SSH tunnel as the logged in user rather than root.
By default, everything inside of the Docker container runs as the user
'root'. Previously, we were setting up the SSH tunnel, we defaulted to
the user inside the VM matching the user inside the Docker
container. However, 'root' is special, and some VM images do not allow
SSH-ing as that user.

For example, this prevented the Datalab image from being able to connect
to a kernel gateway running inside of a Google Dataproc cluster.

This change fixes that be making the username inside of the VM match
the username of the logged-in account (e.g. "<username>@gmail.com").