Support hosts that block websockets.
Previously, Datalab proxied web-socket connections over HTTP, so that
it could run in App Engine Flex (which does not support websockets).

That code was removed when we transitioned to the locally-run setup,
as we did not think that it was necessary anymore.

However, even in the locally-run setup, a user may want to run in
an environment that does not support web-sockets. One such example
is users who want to run Datalab using Google's Cloud Shell Web
Preview feature (https://cloud.google.com/shell/docs/using-web-preview).

This change adds support for such environments by conditionally
proxying web sockets over HTTP based on an environment variable that
can be set by the user.

The code for this is just the old proxy code from the App Engine
Flex version of Datalab, but wrapped in checks based on the new
user setting.

This works for the entirely-locally run version, but additional
work will be required for this to also work with kernel gateways,
as the proxied web socket connections look like cross-domain web
sockets, which are currently blocked.