Fix timing issues in the loading of the app bar
The splitting out of the app bar into a single, reusable HTML file
introduced an unexpected timing issue where the Javascript code served
by the Jupyter notebook server was trying to modify DOM elements which
did not yet exist.

This change fixes that issue by ensuring that the app bar HTML is loaded
before the notebook server Javascript is loaded. Since the loading of
the app bar HTML requires JQuery, this required us to add JQuery as one
of the scripts loaded in each of the HTML templates.