Run ungit inside the datalab container (#1047)
Add ungit inside the datalab container. It runs on port 8083 inside the container, and gets its requests reverse proxied so only the one port used by datalab is open outside the container.
Also added a git icon in the app bar to open ungit in the current location. It finds the current directory path and points ungit to it. Note that ungit will ignore any trailing subdirectories after the git root dir.