Added a release rollback script (#1048)
The rollback relies on a 'previous' variable defined in the config_local.js file. This PR updates the publish script to put the current release as previous before publishing the new one. This way the rollback script can read the previous release version and copy its config_local.js to be the current one, and release its corresponding images as well.