Support specifying the zone via the CLI when creating an instance (#1108)
This fixes #1106 where the `datalab create` command only works if
the `compute/zone` gcloud config is set prior to running the command.

With this fix, either the following will work:

1. Setting the zone via the `--zone` argument to the datalab command
2. Setting the zone via the interactive prompt issued by gcloud

However, if using the second option, the user must make sure that
they select the same zone for both the instance and creating the
disk. That issue will be fixed by a subsequent change to come later.