Automatically detect if websockets need to be proxied.
This change removes the previously used 'PROXY_WEB_SOCKETS' environment
variable, and instead detects on the client whether or not that client
needs to proxy websocket connections via socket.io.

This gives us the benefits that:

1. The user no longer needs to decide ahead of time whether or not
   their instance will rely on websockets.
2. A single instance can both rely on websockets for some clients
   and proxy them for others (e.g. depending on whether an SSH
   tunnel is coming from a Cloud Shell instance or from something else).

For now the logic for this decision is based on whether or not the
hostname ends in '.appspot.com' (indicating that it is hosted by
App Engine which does not support websockets).

However, in the future we may want to make the logic more sophisticated
by first trying to create a websocket connection, and then proxying
them if that fails.