Fix the conflict between the socket.io proxying and ungit
This commit changes socket.io proxying so that it is done
by a separate server listening to port 8084.

That fixes a bug where the socket.io logic on the main server
was interferring with ungit's usage of socket.io.

The reason this fixes that bug is because the socket.io wrapped
server is only invoked after the reverse proxy has determined that
the request should not be proxied to ungit.

This fixes #1125