Add option to shutdown VM from Datalab (#1130)
This PR adds a link in the profile menu to manage the VM hosting Datalab and a button to shut it down. The needed information is stored in environment variables that are computed in the startup script. The VM section appears only if Datalab is hosted inside a google cloud VM.