Datalab test framework (#1115)
This comprises an autotesting script, Selenium driver and framework for handling
exceptions (apparent failures that are expected).

The test.py script can be invoked with a 'vcr' argument in which case HTTP calls
will be recorded to YAML files or played back from those files if they already
exist, which speeds up test execution by an order of magnitude. To use this
feature a special test version of the container is required; it is recommended that
this be used locally only (so there are build and run scripts for this but no stage
or publish scripts). The build-test.sh and run-test.sh files can be used to
do with with Datalab.

To use against a cloud deployment a --profile argumemt must be given so that the
browser opens with the user's authenticated profile.

Tests are specified in a YAML file. Entire notebooks can be ignore, or individual
cells. For even finer control individual MIME-type outputs can be used on a
particular cell.

The test.py and test.js files are generic; all Datalab-specific functionality
is in test_scrubber.js and the YAML file. This means we can potentially turn
this into a generally useful PiP-installable package at some point.