Use VM's zone and name in the GCS backup path (#1137)
This PR uses the VM's zone and name instead of its numeric ID in the GCS backup path to make it easy for the user to find and restore backups. It uses the environment variables added inside the Datalab container for VM_PROJECT, VM_ZONE, and VM_NAME. It also fixes a minor bug with the backup log where it was written prematurely.