Avoid unnecessary dependency upgrades when installing packages using pip.
This change adds the '--upgrade-strategy only-if-needed' flag to all of
our invocations of 'pip install' in the base image.

That flag prevents pip from upgrading a dependency if that dependency
is already satisfied by an installed, older version. That, in turn,
is required to prevent pip from effectively unpinning a version that
we have pinned.

This fixes #1140